requirements.txt
yfinance==0.2.18
matplotlib==3.7.1
crewai-tools==0.1.0
langchain-groq==0.1.0
langchain-openai==0.1.0
langchain-core==0.1.0
python-dotenv==1.0.0
numpy==1.24.3
markdown==3.4.1
pdfkit==1.0.0
PyPDF2==3.0.1
streamlit==1.35.0